simplify agg joiner example
